requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cryptography==42.0.7
defusedxml==0.7.1
Django==5.0.6
django-allauth==0.63.1
django-cors-headers==4.3.1
django-extensions==3.2.3
django-polymorphic==3.1.0
django-rest-auth==0.9.5
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-jwt==1.19.2
drf-pyotp==0.6.0
drf-writable-nested==0.7.0
idna==3.7
mysqlclient==2.2.4
oauthlib==3.2.2
pillow==10.3.0
pycparser==2.22
PyJWT==2.8.0
pyotp==2.9.0
python3-openid==3.2.0
pytz==2024.1
requests==2.32.2
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1